pyproject.toml
[tool.poetry]
name = "etna-cli"
version = "0.2.8"
description = "Command line tool for my school's intranet"
authors = ["matteyeux"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/matteyeux/etna"
repository = "https://github.com/matteyeux/etna"
keywords = ['etna']
classifiers=[
	'Development Status :: 4 - Beta',
	'Intended Audience :: Developers',
	'Intended Audience :: Education',
	'Intended Audience :: System Administrators',
	'Topic :: Utilities',
	'License :: OSI Approved :: MIT License',
	'Programming Language :: Python :: 3.6',
	'Programming Language :: Python :: 3.7',
	'Programming Language :: Python :: 3.8',
]

[tool.poetry.dependencies]
python = "^3.7"
etnawrapper = "^2.5.0"
click = "^7.0"
keyring = "^21.1.0"
taskw = "^1.2.0"
python-gitlab = "^2.0.1"
pendulum = "^2.0.5"
tabulate = "^0.8.7"
click-default-group = "^1.2.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"

[tool.poetry.scripts]
etna = "etna_cli.etna:main"
etna-cli = "etna_cli.etna:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
